% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr}
\alias{alr}
\title{Simplex matrix to additive logratio matrix}
\usage{
alr(X, d)
}
\arguments{
\item{X}{A matrix of compositions in simplex space}

\item{d}{Index of the column to be used as denominator}
}
\value{
An alr matrix of D-1 elements
}
\description{
This function takes a composition matrix X with D elements in simplex space, closes it 1 and converts it
to a matrix in additive logratio space with D-1 elements with column d as the denominator.
}
