% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averager.R
\name{averager}
\alias{averager}
\title{Average to reduce inputs and their noise
This is to reduce the number of counts associated with concentrations.}
\usage{
averager(X, Y)
}
\arguments{
\item{X}{Elemental counts matrix}

\item{Y}{Reference concentrations}
}
\value{
A reduced X and Y matrix
}
\description{
Average to reduce inputs and their noise
This is to reduce the number of counts associated with concentrations.
}
