% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{Calibrate XRF counts to element concentrations
This function takes an XRF counts matrix and reference concentrations
count matrix and outputs alphas and betas.}
\usage{
calib(X, Y, method = "RF")
}
\arguments{
\item{X}{Elemental counts matrix}

\item{Y}{Reference concentrations}

\item{method}{}
}
\value{
A matrix of alphas and betas
}
\description{
Calibrate XRF counts to element concentrations
This function takes an XRF counts matrix and reference concentrations
count matrix and outputs alphas and betas.
}
