% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalr.R
\name{invalr}
\alias{invalr}
\title{Convert back a matrix from additive logratio to simplex space}
\usage{
invalr(X, d)
}
\arguments{
\item{X}{A matrix of compositions in alr}

\item{d}{Index of column to be used as denominator}
}
\value{
A matrix in simplex space with D columns
}
\description{
This functions take a matrix X of compositions of D-1 elements in alr space and converts it
to simplex space with D elements with column d as the denominator.
}
