% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Predict new concentrations based on a trained model}
\usage{
pred(mdl, newX)
}
\arguments{
\item{mdl}{A model object created by xrfcal::calib()}

\item{newX}{An elemental counts matrix to make concentration predictions based on the trained model}
}
\value{
A matrix of predicted elemental concentrations
}
\description{
This function takes a trained calibration model and a matrix of elemental concentrations
and predicts elemental concentrations for the matrix.
}
