% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xrf}
\alias{xrf}
\title{Example XRF counts and reference concentration data}
\format{
An object of class \code{list} of length 2.
}
\usage{
xrf
}
\description{
An example of XRF counts and reference concentrations to be used. \code{xrf$X} is the elemental counts measured by XRF core scanner
and \code{xrf$Y} are reference elemental concentrations measured by ICP.
}
\references{
ADD REFERENCE LATER
}
\keyword{datasets}
