% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr.R
\name{alr}
\alias{alr}
\title{Convert a matrix to additive logratio matrix}
\usage{
alr(X, d)
}
\arguments{
\item{X}{A matrix of compositions}

\item{d}{Index of column to be used as denominator}
}
\value{
An alr matrix of D-1 elements
}
\description{
This functions take a matrix X of compositions of D elements, close it and convert it
to additive logratio space with D-1 elements with column d as the denominator.
}
