% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averager.R
\name{averager}
\alias{averager}
\title{Reduce the dataset to a representative dataset}
\usage{
averager(X, Y)
}
\arguments{
\item{X}{A matrix of XRF counts. Zeros should be replaced beforehand.}

\item{Y}{A matirx of reference concentrations.}
}
\value{
A list with two objects, reduced X, \code{obj[[1]]}, reduced Y, \code{obj[[2]]}
}
\description{
This function takes counts matrix X and concentrations matrix Y and reduce it to a smaller representative dataset by taking the average, first and third quantile
of counts for each unique concentration. This is useful when a unique concentration is associated with many count instances.
}
\examples{
#Replace zeros in X
dl <- apply(xrf$X, 2, agrmt::minnz)
nzX <- zCompositions::multRepl(xrf$X, dl = dl, label = 0)

#Reduce dataset
xrfRed <- averager(nzX,xrf$Y)
Xred <- xrfRed[[1]]
Yred <- xrfRed[[2]]

}
