% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{Calibrate XRF counts to element concentrations
This function takes an XRF counts matrix and reference concentrations
and returns a model object with R2s, RMSEs and  a predictor function for new data.}
\usage{
calib(X, Y, method = "RF", dl = 0, oalr = FALSE, mtry = ncol(X), ntree = 20)
}
\arguments{
\item{X}{Elemental counts matrix}

\item{Y}{Reference concentrations matrix}

\item{method}{Regression model. Either "LRCE", "MLR", "Cubist" or "RF". RF and Cubist generally take longer time to train}

\item{dl}{(Optional) Detection limits to use for zero replacement algorithm. A vector with D elements. If not provided will be estimated.}

\item{oalr}{A boolean parameter. If true the model results and predictor function will be in alr space.}

\item{mtry}{The number of variables for random forest to sample randomly at each split}

\item{ntree}{Number of estimators for random forest}
}
\value{
A model object that includes R2s, RMSEs and a predictor function, pred.
}
\description{
Calibrate XRF counts to element concentrations
This function takes an XRF counts matrix and reference concentrations
and returns a model object with R2s, RMSEs and  a predictor function for new data.
}
\examples{
#Using the example data in the package build a calibration model
mdl <- calib(xrf$X,xrf$Y, method = "MLR")

}
\references{
Insert own publication later

Weltje, G. J., & Tjallingii, R. (2008). Calibration of XRF core scanners for quantitative geochemical logging of sediment cores: Theory and application. Earth and Planetary Science Letters, 274(3), 423-438. https://doi.org/https://doi.org/10.1016/j.epsl.2008.07.054
}
