% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xrf}
\alias{xrf}
\title{An example of XRF counts and reference concentrations to be used}
\format{
An object of class \code{list} of length 2.
}
\usage{
xrf
}
\description{
An example of XRF counts and reference concentrations to be used
}
\keyword{datasets}
